a2a_poc/pyproject.toml
[project]
name = "a2a-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.5",
    "agno[all]>=1.5.5",
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "firecrawl>=2.7.1",
    "google-genai>=1.17.0",
    "googlesearch-python>=1.3.0",
    "httpx>=0.28.1",
    "langchain-cerebras>=0.5.0",
    "langchain-core>=0.3.62",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.7",
    "loguru>=0.7.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "youtube-transcript-api>=1.0.3",
]
